Set values for objects subclassing `mvdf_obj`

This function returns a new object of the same class as `object` with updated values. Use it as a convenient, pipe-able way to set values for objects subclassing `mvdf_obj`.

set_values(
  object,
  mvdf = NULL,
  metadata = NULL,
  appendix = NULL,
  newclass = NULL,
  ...
)

set_mvdf(mvdf, object, metadata = NULL, appendix = NULL, newclass = NULL, ...)

set_metadata(
  metadata,
  object,
  mvdf = NULL,
  appendix = NULL,
  newclass = NULL,
  ...
)

set_appendix(
  appendix,
  object,
  mvdf = NULL,
  metadata = NULL,
  newclass = NULL,
  ...
)

Arguments

object
The object to update.

mvdf
The minimum viable data frame required by the S4 class. If `NULL` (the default), uses the mvdf from `object`.

metadata
The metadata to include in the new object. If `NULL` (the default), uses the metadata from `object`.

appendix
The appendix to include in the new object. If `NULL` (the default), uses the appendix from `object`.

newclass
The class of the object to return. If `NULL` (the default), returns an object of class `class(object)`.

...
Any additional arguments used in the constructor function being called.

Value

An S4 object (of class `newclass` if specified or `class(object)` if not) with updated values.